many more windows fixes
git-svn-id: svn://svn.code.sf.net/p/libsh/code/shrike/branches/win32@1281 afdca40c-03d6-0310-8ede-e9f093b21075

+++ src/Timer.cpp
@@ -15,7 +15,7 @@ float ShTimer::value(void)
   {
   LARGE_INTEGER freq;
   QueryPerformanceFrequency(&freq);
-  return (float)(t.QuadPart)/(freq.QuadPart/1000);
+  return ((float)t.QuadPart)/((float)freq.QuadPart/1000.0);
   }
 
 ShTimer ShTimer::now(void)
